+++ js/configure-provider/constants.js
@@ -1,0 +1,14 @@
+const provider = 'provider';
+const providerUseExistingKey = 'useExistingKey';
+const providerCredsPath = 'credsPath';
+const credsConfirmOverwrite = 'confirmOverwrite';
+
+const inputCredsPath = 'credsPath';
+
+module.exports = {
+  provider,
+  providerUseExistingKey,
+  providerCredsPath,
+  credsConfirmOverwrite,
+  inputCredsPath,
+};